+++ foyer-storage/src/picker/utils.rs
@@ -13,10 +13,8 @@
 // limitations under the License.
 
 use std::{
-    collections::VecDeque,
     fmt::Debug,
     num::NonZeroUsize,
-    ops::Range,
     sync::{
         atomic::{AtomicUsize, Ordering},
         Arc,
@@ -24,14 +22,8 @@ use std::{
     time::Duration,
 };
 
-use foyer_common::strict_assert;
-use itertools::Itertools;
-
-use super::{AdmissionPicker, EvictionInfo, EvictionPicker, Pick, ReinsertionPicker};
-use crate::{
-    io::{device::RegionId, throttle::IoThrottler},
-    statistics::Statistics,
-};
+use super::{AdmissionPicker, Pick, ReinsertionPicker};
+use crate::{io::throttle::IoThrottler, statistics::Statistics};
 
 /// Only admit on all chained admission pickers pick.
 #[derive(Debug, Default, Clone)]
@@ -191,265 +183,5 @@ impl IoThrottlerPicker {
 impl AdmissionPicker for IoThrottlerPicker {
     fn pick(&self, stats: &Arc<Statistics>, _: u64) -> Pick {
         self.pick_inner(stats)
-    }
-}
-
-/// A picker that pick region to eviction with a FIFO behavior.
-#[derive(Debug)]
-pub struct FifoPicker {
-    queue: VecDeque<RegionId>,
-    regions: usize,
-    probations: usize,
-    probation_ratio: f64,
-}
-
-impl Default for FifoPicker {
-    fn default() -> Self {
-        Self::new(0.1)
-    }
-}
-
-impl FifoPicker {
-    /// Create a new [`FifoPicker`] with the given `probation_ratio` (0.0 ~ 1.0).
-    ///
-    /// The `probation_ratio` is the ratio of regions that will be marked as probation.
-    /// The regions that are marked as probation will be evicted first.
-    ///
-    /// The default value is 0.1.
-    pub fn new(probation_ratio: f64) -> Self {
-        assert!(
-            (0.0..=1.0).contains(&probation_ratio),
-            "probation ratio {probation_ratio} must be in [0.0, 1.0]"
-        );
-        Self {
-            queue: VecDeque::new(),
-            regions: 0,
-            probations: 0,
-            probation_ratio,
-        }
-    }
-}
-
-impl FifoPicker {
-    fn mark_probation(&self, info: EvictionInfo<'_>) {
-        let marks = self.probations.saturating_sub(info.clean);
-        self.queue.iter().take(marks).for_each(|rid| {
-            if info.evictable.contains(rid) {
-                info.regions[*rid as usize]
-                    .statistics()
-                    .probation
-                    .store(true, Ordering::Relaxed);
-                tracing::trace!(rid, "[fifo picker]: mark probation");
-            }
-        });
-    }
-}
-
-impl EvictionPicker for FifoPicker {
-    fn init(&mut self, regions: Range<RegionId>, _: usize) {
-        self.regions = regions.len();
-        let probations = (self.regions as f64 * self.probation_ratio).floor() as usize;
-        self.probations = probations.clamp(0, self.regions);
-    }
-
-    fn pick(&mut self, info: EvictionInfo<'_>) -> Option<RegionId> {
-        tracing::trace!(evictable = ?info.evictable, clean = info.clean, "[fifo picker]: pick");
-        self.mark_probation(info);
-        let candidate = self.queue.front().copied();
-        tracing::trace!("[fifo picker]: pick {candidate:?}");
-        candidate
-    }
-
-    fn on_region_evictable(&mut self, _: EvictionInfo<'_>, region: RegionId) {
-        tracing::trace!("[fifo picker]: {region} is evictable");
-        self.queue.push_back(region);
-    }
-
-    fn on_region_evict(&mut self, _: EvictionInfo<'_>, region: RegionId) {
-        tracing::trace!("[fifo picker]: {region} is evicted");
-        let index = self.queue.iter().position(|r| r == &region).unwrap();
-        self.queue.remove(index);
-    }
-}
-
-/// Evict the region with the largest invalid data ratio.
-///
-/// If the largest invalid data ratio is less than the threshold, no region will be picked.
-#[derive(Debug)]
-pub struct InvalidRatioPicker {
-    threshold: f64,
-    region_size: usize,
-}
-
-impl InvalidRatioPicker {
-    /// Create [`InvalidRatioPicker`] with the given `threshold` (0.0 ~ 1.0).
-    pub fn new(threshold: f64) -> Self {
-        let ratio = threshold.clamp(0.0, 1.0);
-        Self {
-            threshold: ratio,
-            region_size: 0,
-        }
-    }
-}
-
-impl EvictionPicker for InvalidRatioPicker {
-    fn init(&mut self, _: Range<RegionId>, region_size: usize) {
-        self.region_size = region_size;
-    }
-
-    fn pick(&mut self, info: EvictionInfo<'_>) -> Option<RegionId> {
-        strict_assert!(self.region_size > 0);
-
-        let mut data = info
-            .evictable
-            .iter()
-            .map(|rid| {
-                (
-                    *rid,
-                    info.regions[*rid as usize].statistics().invalid.load(Ordering::Relaxed),
-                )
-            })
-            .collect_vec();
-        data.sort_by_key(|(_, invalid)| *invalid);
-
-        let (rid, invalid) = data.last().copied()?;
-        if (invalid as f64 / self.region_size as f64) < self.threshold {
-            return None;
-        }
-        tracing::trace!("[invalid ratio picker]: pick {rid:?}");
-        Some(rid)
-    }
-
-    fn on_region_evictable(&mut self, _: EvictionInfo<'_>, region: RegionId) {
-        tracing::trace!("[invalid ratio picker]: {region} is evictable");
-    }
-
-    fn on_region_evict(&mut self, _: EvictionInfo<'_>, region: RegionId) {
-        tracing::trace!("[invalid ratio picker]: {region} is evicted");
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashSet;
-
-    use super::*;
-    use crate::{io::engine::noop::NoopIoEngine, region::Region};
-
-    #[test_log::test]
-    fn test_fifo_picker() {
-        let mut picker = FifoPicker::new(0.1);
-
-        let regions = (0..10)
-            .map(|rid| Region::new_for_test(rid, Arc::<NoopIoEngine>::default()))
-            .collect_vec();
-        let mut evictable = HashSet::new();
-
-        fn info<'a>(regions: &'a [Region], evictable: &'a HashSet<RegionId>, dirty: usize) -> EvictionInfo<'a> {
-            EvictionInfo {
-                regions,
-                evictable,
-                clean: regions.len() - evictable.len() - dirty,
-            }
-        }
-
-        fn check(regions: &[Region], probations: impl IntoIterator<Item = RegionId>) {
-            let probations = probations.into_iter().collect::<HashSet<_>>();
-            for rid in 0..regions.len() as RegionId {
-                if probations.contains(&rid) {
-                    assert!(
-                        regions[rid as usize].statistics().probation.load(Ordering::Relaxed),
-                        "probations {probations:?}, assert {rid} is probation failed"
-                    );
-                } else {
-                    assert!(
-                        !regions[rid as usize].statistics().probation.load(Ordering::Relaxed),
-                        "probations {probations:?}, assert {rid} is not probation failed"
-                    );
-                }
-            }
-        }
-
-        picker.init(0..10, 0);
-
-        // mark 0..10 evictable in order
-        (0..10).for_each(|i| {
-            evictable.insert(i as _);
-            picker.on_region_evictable(info(&regions, &evictable, 0), i);
-        });
-
-        // evict 0, mark 0 probation.
-        assert_eq!(picker.pick(info(&regions, &evictable, 0)), Some(0));
-        check(&regions, [0]);
-        evictable.remove(&0);
-        picker.on_region_evict(info(&regions, &evictable, 1), 0);
-
-        // evict 1, with 1 dirty region, mark 1 probation.
-        assert_eq!(picker.pick(info(&regions, &evictable, 1)), Some(1));
-        check(&regions, [0, 1]);
-        evictable.remove(&1);
-        picker.on_region_evict(info(&regions, &evictable, 2), 1);
-
-        // evict 2 with 1 dirty region, mark no region probation.
-        assert_eq!(picker.pick(info(&regions, &evictable, 1)), Some(2));
-        check(&regions, [0, 1]);
-        evictable.remove(&2);
-        picker.on_region_evict(info(&regions, &evictable, 2), 2);
-
-        picker.on_region_evict(info(&regions, &evictable, 3), 3);
-        evictable.remove(&3);
-        picker.on_region_evict(info(&regions, &evictable, 4), 5);
-        evictable.remove(&5);
-        picker.on_region_evict(info(&regions, &evictable, 5), 7);
-        evictable.remove(&7);
-        picker.on_region_evict(info(&regions, &evictable, 6), 9);
-        evictable.remove(&9);
-
-        picker.on_region_evict(info(&regions, &evictable, 7), 4);
-        evictable.remove(&4);
-        picker.on_region_evict(info(&regions, &evictable, 8), 6);
-        evictable.remove(&6);
-        picker.on_region_evict(info(&regions, &evictable, 9), 8);
-        evictable.remove(&8);
-    }
-
-    #[test]
-    fn test_invalid_ratio_picker() {
-        let mut picker = InvalidRatioPicker::new(0.5);
-        picker.init(0..10, 10);
-
-        let regions = (0..10)
-            .map(|rid| Region::new_for_test(rid, Arc::<NoopIoEngine>::default()))
-            .collect_vec();
-        let mut evictable = HashSet::new();
-
-        (0..10).for_each(|i| {
-            regions[i].statistics().invalid.fetch_add(i as _, Ordering::Relaxed);
-            evictable.insert(i as RegionId);
-        });
-
-        fn info<'a>(regions: &'a [Region], evictable: &'a HashSet<RegionId>) -> EvictionInfo<'a> {
-            EvictionInfo {
-                regions,
-                evictable,
-                clean: regions.len() - evictable.len(),
-            }
-        }
-
-        assert_eq!(picker.pick(info(&regions, &evictable)), Some(9));
-
-        assert_eq!(picker.pick(info(&regions, &evictable)), Some(9));
-        evictable.remove(&9);
-        assert_eq!(picker.pick(info(&regions, &evictable)), Some(8));
-        evictable.remove(&8);
-        assert_eq!(picker.pick(info(&regions, &evictable)), Some(7));
-        evictable.remove(&7);
-        assert_eq!(picker.pick(info(&regions, &evictable)), Some(6));
-        evictable.remove(&6);
-        assert_eq!(picker.pick(info(&regions, &evictable)), Some(5));
-        evictable.remove(&5);
-
-        assert_eq!(picker.pick(info(&regions, &evictable)), None);
-        assert_eq!(picker.pick(info(&regions, &evictable)), None);
     }
 }
